pyproject.toml
[project]
name = "ps-model-selection-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.4",
    "pdf2image>=1.17.0",
    "pillow>=11.2.0",
    "pydantic-ai-slim[openai]>=0.2.16",
    "pytest>=8.4.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "textstat>=0.7.0",
    "nltk>=3.8.0",
]

[project.scripts]
dev = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
test = "pytest"
